requirements.txt
Django==1.8.6
Markdown==2.6.2
Pillow==3.2.0
argparse==1.2.1
django-modeltranslation==0.11
django-rosetta==0.7.6
django-taggit==0.17.1
easy-thumbnails==2.3
microsofttranslator==0.5
polib==1.0.7
psycopg2==2.6.1
pytz==2016.3
requests==2.9.1
six==1.10.0
wsgiref==0.1.2
